# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Astra Gateway API
  version: 1.0.0
  description: |
    Enterprise AI Gateway Platform — unified access to multiple LLM providers with intelligent routing,
    governance, observability, Agent-to-Agent (A2A) communication, and Model Context Protocol (MCP) support.

    **Services & Ports (local dev)**
    | Service             | Port |
    |---------------------|------|
    | Gateway Service     | 8080 |
    | MCP Service         | 8081 |
    | A2A Service         | 8082 |
    | Auth Service        | 8083 |
    | Routing Engine      | 8084 |
    | Governance Engine   | 8085 |
    | Observability       | 8086 |
  contact:
    name: Astra Gateway Support
    url: https://docs.astragateway.io
    email: support@astragateway.io
  license:
    name: Proprietary
    url: https://astragateway.io/license
servers:
  - url: https://api.astragateway.io/v1
    description: Production (Gateway)
  - url: https://staging.astragateway.io/v1
    description: Staging (Gateway)
  - url: http://localhost:8080/v1
    description: Local — Gateway Service
  - url: http://localhost:8081/v1
    description: Local — MCP Service
  - url: http://localhost:8082/v1
    description: Local — A2A Service
  - url: http://localhost:8083/v1
    description: Local — Auth Service
  - url: http://localhost:8084/v1
    description: Local — Routing Engine
  - url: http://localhost:8085/v1
    description: Local — Governance Engine
  - url: http://localhost:8086/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /routing/decide:
    post:
      tags:
        - Routing
      summary: Get a routing decision
      description: |
        Returns the selected provider and a full fallback chain for a given model and strategy.

        **Strategies:**
        - `cost` — cheapest provider first (Cohere → OpenAI → Anthropic)
        - `latency` — fastest provider first based on runtime averages
        - `quality` — highest quality-score provider first (Anthropic → OpenAI → Cohere)

        Routing decisions are cached in Redis for 60 seconds.
      operationId: decideRouting
      servers:
        - url: http://localhost:8084/v1
          description: Local — Routing Engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutingDecisionRequest'
            examples:
              latency:
                summary: Latency-optimised routing
                value:
                  model: claude-sonnet-4-6
                  strategy: latency
              cost:
                summary: Cost-optimised routing
                value:
                  model: gpt-4o
                  strategy: cost
      responses:
        '200':
          description: Routing decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingDecisionResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token verified by the Auth Service (`POST /auth/verify`)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API key validated by the Auth Service (`POST /auth/api-key/validate`).
        Format: `Bearer sk-astra-<key>`
  schemas:
...
    RoutingDecisionRequest:
      type: object
      properties:
        model:
          type: string
          example: claude-sonnet-4-6
        strategy:
          type: string
          enum: [cost, latency, quality]
          default: latency
    RoutingDecisionResponse:
      type: object
      properties:
        selected_provider:
          type: string
          example: anthropic
        model:
          type: string
        strategy:
          type: string
        fallback_chain:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              priority:
                type: integer
        constraints:
          type: object
          properties:
            max_tokens:
              type: integer
            timeout_ms:
              type: integer
        timestamp:
          type: integer
...